form_template_app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.17
djangorestframework==3.15.2
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.2
tinydb==4.8.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
